rcm_project/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
numpy==1.26.4
pandas==2.2.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
asgiref             3.7.2          
joblib              1.3.2                
pip                 24.0                
scikit-learn        1.4.1.post1
scipy               1.12.0     
setuptools          65.5.0                
threadpoolctl       3.3.0      
     
